+++ scripts/babel-6/__tests__/rewriteModules-test.js
@@ -1,10 +1,11 @@
 /**
- * Copyright (c) 2014, Facebook, Inc. All rights reserved.
+ * Copyright (c) 2015, Facebook, Inc. All rights reserved.
  *
  * This source code is licensed under the BSD-style license found in the
  * LICENSE file in the root directory of this source tree. An additional grant
  * of patent rights can be found in the PATENTS file in the same directory.
  */
+
 'use strict';
 
 jest.dontMock('babel-core');
@@ -21,17 +22,18 @@ describe('rewrite-modules', function() {
   describe('rewriteModules', function() {
     describe('opts._moduleMap', function() {
       it('should replace the prefix on normal requires', function() {
-        let result = babel.transform('require(\'test\');',
+        let result = babel.transform(
+          'require(\'test\');',
           {
-            plugins: [[rewriteModules, {_moduleMap: {'test': 'test/test'}}]]
-          });
+            plugins: [[rewriteModules, {_moduleMap: {'test': 'test/test'}}]],
+          }
+        );
 
         expect(result.code).toEqual('require(\'test/test\');');
       });
 
-      it('should work for jest. mock, dontMock ' +
-        ' and genMockFromModule', function() {
-        const code = (function test() {
+      it('should work for jest. mock, dontMock and genMockFromModule', function() {
+        const code = `function test() {
           'use strict';
 
           jest.mock('foo');
@@ -40,9 +42,9 @@ describe('rewrite-modules', function() {
 
           var foo = require('foo');
           var actualFoo = require.requireActual('foo');
-        }).toString();
+        }`;
 
-        const expected = (function test() {
+        const expected = `function test() {
           'use strict';
 
           jest.mock('foo/foo');
@@ -51,24 +53,28 @@ describe('rewrite-modules', function() {
 
           var foo = require('foo/foo');
           var actualFoo = require.requireActual('foo/foo');
-        }).toString();
+        }`;
 
-        const rewritePlugin = [rewriteModules, {
-          _moduleMap: {
-            'foo': 'foo/foo',
-            'bar': 'bar/bar',
-            'baz': 'baz/baz'
+        const rewritePlugin = [
+          rewriteModules, {
+            _moduleMap: {
+              'foo': 'foo/foo',
+              'bar': 'bar/bar',
+              'baz': 'baz/baz',
+            },
+          },
+        ];
+
+        let result = babel.transform(
+          code,
+          {
+            plugins: [rewritePlugin],
           }
-        }];
-
-        let result = babel.transform(code, {
-          plugins: [rewritePlugin]
-        });
+        );
 
         expect(normalizeResults(result.code))
           .toEqual(normalizeResults(expected));
       });
     });
   });
-
 });